updated db tables for current dev code.

+++ sql/tables/bookmarks.sql
@@ -23,19 +23,19 @@ DROP TABLE IF EXISTS `bookmarks`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `bookmarks` (
-  `bookmarkID` int(10)  NOT NULL AUTO_INCREMENT,
-  `ownerID` int(10)  NOT NULL DEFAULT '0',
-  `itemID` int(10)  NOT NULL DEFAULT '0',
-  `typeID` int(10)  NOT NULL DEFAULT '0',
+  `bookmarkID` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `ownerID` int(10) unsigned NOT NULL DEFAULT '0',
+  `itemID` int(10) unsigned NOT NULL DEFAULT '0',
+  `typeID` int(10) unsigned NOT NULL DEFAULT '0',
   `memo` varchar(85) NOT NULL DEFAULT '',
-  `created` bigint(20)  NOT NULL DEFAULT '0',
+  `created` bigint(20) unsigned NOT NULL DEFAULT '0',
   `x` double NOT NULL DEFAULT '0',
   `y` double NOT NULL DEFAULT '0',
   `z` double NOT NULL DEFAULT '0',
-  `locationID` int(10)  NOT NULL DEFAULT '0',
+  `locationID` int(10) unsigned NOT NULL DEFAULT '0',
   `note` varchar(85) NOT NULL DEFAULT '',
-  `creatorID` int(10)  NOT NULL DEFAULT '0',
-  `folderID` int(10)  NOT NULL DEFAULT '0',
+  `creatorID` int(10) unsigned NOT NULL DEFAULT '0',
+  `folderID` smallint(6) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`bookmarkID`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 /*!40101 SET character_set_client = @saved_cs_client */;
@@ -58,4 +58,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2017-08-01 18:36:20
+-- Dump completed on 2018-01-16 19:11:33
